+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Migrated codebase to TypeScript.
   used to control complexity threshold objects must exceed to be formatted over
   multiple lines.
 - Show when formatted value includes multiple references to the same object.
+- Show when formatted value is a proxy (in Node.js).
 - Formatters for primitive wrapper objects: `Boolean`, `Number`, and `String`.
 - Formatter for `Error`s.
 - Include `Symbol.toStringTag` in object's name.